+++ engine-kqueue/src/macosMain/kotlin/io/github/fukusaka/keel/engine/kqueue/KqueueEventLoopGroup.kt
@@ -1,5 +1,6 @@
 package io.github.fukusaka.keel.engine.kqueue
 
+import io.github.fukusaka.keel.logging.Logger
 import kotlin.concurrent.AtomicInt
 
 /**
@@ -14,10 +15,11 @@ import kotlin.concurrent.AtomicInt
  * Array + AtomicInt round-robin.
  *
  * @param size Number of EventLoop threads. Must be >= 1.
+ * @param logger Logger for each EventLoop in the group.
  */
-internal class KqueueEventLoopGroup(size: Int) {
+internal class KqueueEventLoopGroup(size: Int, logger: Logger) {
 
-    private val loops = Array(size) { KqueueEventLoop() }
+    private val loops = Array(size) { KqueueEventLoop(logger) }
     private val index = AtomicInt(0)
 
     /** Starts all EventLoop threads. */